Changed the fit bonus test a bit
