--- ElectronicObserverCoreTests/FitBonusTest.cs
@@ -1,349 +1,0 @@
-﻿using System.Collections.Generic;
-using ElectronicObserver.Utility.Data;
-using ElectronicObserverTypes;
-using ElectronicObserverTypes.Mocks;
-using ElectronicObserverTypes.Serialization.FitBonus;
-using Xunit;
-
-namespace ElectronicObserverCoreTests;
-
-[Collection(DatabaseCollection.Name)]
-public class FitBonusTest
-{
-	private DatabaseFixture Db { get; }
-
-	private ElectronicObserver.Data.Translation.FitBonusData BonusData { get; }
-
-	public FitBonusTest(DatabaseFixture db)
-	{
-		Db = db;
-		BonusData = new();
-	}
-
-	[Fact]
-	public void FitBonusTest1()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData hachijou = new ShipDataMock(Db.MasterShips[ShipId.HachijouKai])
-		{
-			Level = 175,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_12cmSingleHighangleGunModelE]),
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_12cmSingleHighangleGunModelE])
-				{
-					Level = 10
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.RadarSmall_GFCSMk_37]),
-			}
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 2,
-			Evasion = 3 + 3 + (2 * 2),
-			AntiAir = 2 + (2 * 2),
-			ASW = 1 * 2,
-			Accuracy = 0,
-			Armor = 0,
-			LOS = 0,
-			Range = 0,
-			Torpedo = 0,
-		};
-
-		FitBonusValue finalBonus = hachijou.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest2()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData yikikaze = new ShipDataMock(Db.MasterShips[ShipId.YukikazeKaiNi])
-		{
-			Level = 151,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_10cmTwinHighangleGun_AAFD]),
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_10cmTwinHighangleGun_AAFD])
-				{
-					Level = 4
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.Torpedo_61cmQuadruple_OxygenTorpedoMountLateModel])
-				{
-					Level = 3
-				},
-			}
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 5,
-			Torpedo = 2,
-			AntiAir = 3,
-			Accuracy = 0,
-			ASW = 0,
-			Armor = 0,
-			Evasion = 2 + 1,
-			LOS = 0,
-			Range = 0,
-		};
-
-		FitBonusValue finalBonus = yikikaze.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest3()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData yikikaze = new ShipDataMock(Db.MasterShips[ShipId.YukikazeKaiNi])
-		{
-			Level = 151,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_10cmTwinHighangleGun_AAFD])
-				{
-					Level = 10
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_10cmTwinHighangleGun_AAFD])
-				{
-					Level = 10
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.Torpedo_61cmQuadruple_OxygenTorpedoMountLateModel])
-				{
-					Level = 10
-				},
-			},
-			ExpansionSlotInstance = new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SurfaceShipPersonnel_TorpedoSquadronSkilledLookouts])
-			{
-				Level = 4
-			},
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 5 + 5 + 1 + 3,
-			Torpedo = 3 + 4,
-			AntiAir = 3 + 3,
-			Accuracy = 0,
-			ASW = 2,
-			Armor = 0,
-			Evasion = 2 + 2 + 1 + 3,
-			LOS = 1,
-			Range = 0,
-		};
-
-		FitBonusValue finalBonus = yikikaze.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest4()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData yamato = new ShipDataMock(Db.MasterShips[ShipId.YamatoKaiNiJuu])
-		{
-			Level = 99,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunLarge_51cmTwinGun])
-				{
-					Level = 2
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunLarge_Prototype51cmTwinGun])
-				{
-					Level = 1
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SeaplaneRecon_Type0ReconSeaplaneModel11B_Skilled]),
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.APShell_Type1ArmorPiercingShellKai])
-				{
-					Level = 4
-				},
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SecondaryGun_15_5cmTripleSecondaryGunMountKaiNi])
-				{
-					Level = 8
-				},
-			},
-			ExpansionSlotInstance = new EquipmentDataMock(Db.MasterEquipment[EquipmentId.RadarLarge_15mDuplexRangefinder_Type21AirRADARKaiNi]),
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 1 + 1 + 2 + 2 + 1,
-			Torpedo = 0,
-			AntiAir = 2 + 1,
-			Accuracy = 1 + 1 + 3 + 3,
-			ASW = 0,
-			Armor = 0,
-			Evasion = 2 + 1,
-			LOS = 0,
-			Range = 0,
-		};
-
-		FitBonusValue finalBonus = yamato.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest5()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData kazagumo = new ShipDataMock(Db.MasterShips[ShipId.KazagumoKaiNi])
-		{
-			Level = 180,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.RadarSmall_PassiveRadiolocator_E27_Type22SurfaceRadarKai4_CalibratedLateModel])
-			}
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 1,
-			Accuracy = 2,
-			Evasion = 1,
-			LOS = 1
-		};
-
-		FitBonusValue finalBonus = kazagumo.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-
-		kazagumo.SlotInstance.Add(
-			new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_12_7cmTwinGunModelDKaiNi]));
-
-		expectedBonus.Firepower += 6;
-		expectedBonus.Torpedo += 6;
-		expectedBonus.Evasion += 4;
-
-		finalBonus = kazagumo.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest6()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData kazagumo = new ShipDataMock(Db.MasterShips[ShipId.KazagumoKaiNi])
-		{
-			Level = 180,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.RadarSmall_PassiveRadiolocator_E27_Type22SurfaceRadarKai4_CalibratedLateModel])
-			}
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue()
-		{
-			Firepower = 1,
-			Accuracy = 2,
-			Evasion = 1,
-			LOS = 1
-		};
-
-		FitBonusValue finalBonus = kazagumo.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-
-		kazagumo.SlotInstance.Add(
-			new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_12_7cmTwinGunModelDKaiNi]) { UpgradeLevel = UpgradeLevel.Three });
-
-		expectedBonus.Firepower += 8;
-		expectedBonus.Torpedo += 6;
-		expectedBonus.Accuracy += 2;
-		expectedBonus.Evasion += 4;
-
-		finalBonus = kazagumo.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-
-		kazagumo.SlotInstance.Add(
-			new EquipmentDataMock(Db.MasterEquipment[EquipmentId.MainGunSmall_12_7cmTwinGunModelDKaiSan]) { UpgradeLevel = UpgradeLevel.Four });
-
-		expectedBonus.Firepower += 6;
-		expectedBonus.Torpedo += 4;
-		expectedBonus.AntiAir += 3;
-		expectedBonus.Accuracy += 3;
-		expectedBonus.Evasion += 3;
-
-		finalBonus = kazagumo.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-
-	[Fact]
-	public void FitBonusTest7()
-	{
-		Assert.NotEmpty(BonusData.FitBonusList);
-
-		IShipData kuma = new ShipDataMock(Db.MasterShips[ShipId.KumaKai])
-		{
-			Level = 180,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SeaplaneRecon_Walrus])
-			}
-		};
-
-		FitBonusValue expectedBonus = new FitBonusValue();
-
-		FitBonusValue finalBonus = kuma.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-
-		IShipData sheffield = new ShipDataMock(Db.MasterShips[ShipId.SheffieldKai])
-		{
-			Level = 180,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SeaplaneRecon_Walrus])
-			}
-		};
-
-		expectedBonus = new FitBonusValue()
-		{
-			Firepower = 2,
-			ASW = 3,
-			LOS = 2,
-			Evasion = 2
-		};
-
-		finalBonus = sheffield.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-
-		IShipData rodney = new ShipDataMock(Db.MasterShips[ShipId.RodneyKai])
-		{
-			Level = 180,
-			SlotInstance = new List<IEquipmentData?>
-			{
-				new EquipmentDataMock(Db.MasterEquipment[EquipmentId.SeaplaneRecon_Walrus])
-			}
-		};
-
-		expectedBonus = new FitBonusValue()
-		{
-			Firepower = 6,
-			ASW = 3,
-			Accuracy = 2,
-			LOS = 5,
-			Evasion = 4
-		};
-
-		finalBonus = rodney.GetTheoricalFitBonus(BonusData.FitBonusList);
-
-		Assert.Equal(expectedBonus, finalBonus);
-	}
-}
